+++ src/test/java/org/jabref/logic/shared/ConnectorTest.java
@@ -17,15 +17,6 @@ public class ConnectorTest {
     }
 
     public static DBMSConnectionProperties getTestConnectionProperties(DBMSType dbmsType) {
-        switch (dbmsType) {
-            case MYSQL:
-                return new DBMSConnectionPropertiesBuilder().setType(dbmsType).setHost("127.0.0.1").setPort(3800).setDatabase("jabref").setUser("root").setPassword("root").setUseSSL(false).setAllowPublicKeyRetrieval(true).createDBMSConnectionProperties();
-            case POSTGRESQL:
-                return new DBMSConnectionPropertiesBuilder().setType(dbmsType).setHost("localhost").setPort(dbmsType.getDefaultPort()).setDatabase("postgres").setUser("postgres").setPassword("postgres").setUseSSL(false).createDBMSConnectionProperties();
-            case ORACLE:
-                return new DBMSConnectionPropertiesBuilder().setType(dbmsType).setHost("localhost").setPort(32118).setDatabase("jabref").setUser("jabref").setPassword("jabref").setUseSSL(false).createDBMSConnectionProperties();
-            default:
-                return new DBMSConnectionPropertiesBuilder().createDBMSConnectionProperties();
-        }
+        return new DBMSConnectionPropertiesBuilder().setType(dbmsType).setHost("localhost").setPort(dbmsType.getDefaultPort()).setDatabase("postgres").setUser("postgres").setPassword("postgres").setUseSSL(false).createDBMSConnectionProperties();
     }
 }