+++ CHANGELOG.md
@@ -89,7 +89,7 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 - We removed support for case-sensitive and exact search. [#11542](https://github.com/JabRef/jabref/pull/11542)
 - We removed the description of search strings. [#11542](https://github.com/JabRef/jabref/pull/11542)
 - We removed support for importing using the SilverPlatterImporter (`Record INSPEC`). [#11576](https://github.com/JabRef/jabref/pull/11576)
-
+- We removed support for MySQL/MariaDB and Oracle.